+++ src/app/pedido/pedido.component.ts
@@ -1,3 +1,4 @@
+import { ToastrService } from 'ngx-toastr';
 import { element } from 'protractor';
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { PedidoService } from '../services/pedido/pedido.service';
@@ -18,11 +19,11 @@ export class PedidoComponent implements OnInit {
   pedidoList: MatTableDataSource<Pedido>;
   errorMsg: String;
 
-  displayedColumns: string[] = ['id', 'cotacao', 'fornecedor', 'produto', 'quantidade', 'data', 'total', 'action'];
+  displayedColumns: string[] = ['id', 'cotacao', 'fornecedor', 'produto', 'quantidade', 'data', 'status', 'total', 'action'];
 
   @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
 
-  constructor(private pedidoService: PedidoService, public dialog: MatDialog) { }
+  constructor(private pedidoService: PedidoService, public dialog: MatDialog, private toast: ToastrService) { }
 
   ngOnInit() {
     this.getPedidos();
@@ -45,26 +46,62 @@ export class PedidoComponent implements OnInit {
   }
 
   openDialog(element): void {
-    console.log(element);
-    const dialogRef = this.dialog.open(PedidoDialogueComponent, {
-      data: {
-        width: 'auto',
-        height: 'auto',
-        element
-      }
-    });
+    if (element == null || element.status === 'Aguardando transportadora') {
+      const dialogRef = this.dialog.open(PedidoDialogueComponent, {
+        data: {
+          width: 'auto',
+          height: 'auto',
+          element
+        }
+      });
+      dialogRef.afterClosed().subscribe(
+        data => {
+          this.getPedidos();
+        }
+      );
+    }
+    else if (element.status === 'Cancelado') {
+      this.toast.error('Não é possível editar o pedido', 'Pedido cancelado');
+    }
+    else if (element.status === 'Entregue') {
+      this.toast.error('Não é possível editar o pedido', 'Pedido já foi entregue');
+    }
+  }
 
-    dialogRef.afterClosed().subscribe(
-      data => {
-        this.getPedidos();
-      }
-    );
+  updateStatus(pedido: Pedido) {
+    if (pedido.status === 'Entregue') {
+      this.toast.show('Pedido já foi entregue');
+    } else {
+      this.pedidoService.updateStatus(pedido).subscribe(
+        data => {
+          this.toast.success('Status do pedido atualizado');
+          this.getPedidos();
+        });
+    }
+  }
+
+  cancelarPedido(pedido: Pedido) {
+    if (pedido.status === 'Cancelado' || pedido.status === 'Entregue') {
+      this.toast.error('Impossível excluir pedidos com status ' + pedido.status);
+    }
+    else {
+      this.pedidoService.cancelarPedido(pedido).subscribe(
+        data => {
+          this.toast.success('Pedido cancelado');
+          this.getPedidos();
+        });
+    }
   }
 
   deletar(pedido: Pedido) {
-    this.pedidoService.delete(pedido).subscribe(
-      data => {
-        this.getPedidos();
-      });
+    if (pedido.status === 'Cancelado' || pedido.status === 'Entregue') {
+      this.toast.error('Não é possível excluir pedido com status ' + pedido.status, 'Erro');
+    } else {
+      this.pedidoService.delete(pedido).subscribe(
+        data => {
+          this.toast.success('Pedido excluído');
+          this.getPedidos();
+        });
+    }
   }
 }
